Draw key points

+++ KMLPipePy/operations/pose2D.py
@@ -1,5 +1,5 @@
 from KMLPipePy.CVNodeProcess import CVNodeProcess
-from KMLPipePy.base_structs import DataType, Keypoint2D, KPFrame
+from KMLPipePy.types import Keypoint2D, KPFrame
 
 import tensorflow_hub as hub
 import tensorflow as tf
@@ -34,7 +34,7 @@ class Pose2D(CVNodeProcess):
         image = self.process_image(image)
         res = self.movenet(image)
 
-        out = KPFrame(keypoints=[Keypoint2D(x = p[1].numpy() * width, y = p[0].numpy() * height, score = p[2].numpy(), name = LABELS[i])
+        out = KPFrame(keypoints=[Keypoint2D(x = int(p[1].numpy() * width), y = int(p[0].numpy() * height), score = int(p[2].numpy()), name = LABELS[i])
                 for i, p in enumerate(res['output_0'][0][0])])
 
         self.vars["output-0"] = out
